refactor: move memberTypeId
